tried to change colours

+++ srcs/7_image.c
@@ -6,7 +6,7 @@
 /*   By: deelliot <deelliot@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/05 15:34:31 by deelliot          #+#    #+#             */
-/*   Updated: 2022/08/23 12:09:24 by deelliot         ###   ########.fr       */
+/*   Updated: 2022/08/23 12:11:13 by deelliot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ void	img_pixel_put(t_img *img, int x, int y, t_rgb *colour)
 	if (x > 0 && x < WIDTH && y > 0 && y < HEIGHT)
 	{
 		pix = (y * img->length) + (x * img->bpp / 8);
-		img->addr[pix] = (int)colour->r;
-		img->addr[pix + 1] = (int)colour->g;
-		img->addr[pix + 2] = (int)colour->b;
+		img->addr[pix] = colour->r;
+		img->addr[pix + 1] = colour->g;
+		img->addr[pix + 2] = colour->b;
 	}
 }
